+++ framework/log/FileTarget.php
@@ -20,6 +20,8 @@ use yii\helpers\FileHelper;
  * files are moved backwards by one place, i.e., '.2' to '.3', '.1' to '.2', and so on.
  * The property [[maxLogFiles]] specifies how many history files to keep.
  *
+ * Since 2.0.46 rotation of the files is done only by copy and property `rotateByCopy` is deprecated.
+ *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
@@ -60,15 +62,15 @@ class FileTarget extends Target
     public $dirMode = 0775;
     /**
      * @var bool Whether to rotate log files by copy and truncate in contrast to rotation by
-     * renaming files. Defaults to `true` to be more compatible with log tailers and is windows
+     * renaming files. Defaults to `true` to be more compatible with log tailers and windows
      * systems which do not play well with rename on open files. Rotation by renaming however is
      * a bit faster.
      *
      * The problem with windows systems where the [rename()](https://www.php.net/manual/en/function.rename.php)
      * function does not work with files that are opened by some process is described in a
      * [comment by Martin Pelletier](https://www.php.net/manual/en/function.rename.php#102274) in
-     * the PHP documentation. By setting rotateByCopy to `true` you can work
-     * around this problem.
+     * the PHP documentation. By setting rotateByCopy to `true` you can work around this.
+     * @deprecated since 2.0.46 and setting it to false has no effect anymore since rotating is done only by copy
      */
     public $rotateByCopy = true;
 
@@ -117,9 +119,9 @@ class FileTarget extends Target
             clearstatcache();
         }
         if ($this->enableRotation && @filesize($this->logFile) > $this->maxFileSize * 1024) {
+            $this->rotateFiles();
             @flock($fp, LOCK_UN);
             @fclose($fp);
-            $this->rotateFiles();
             $writeResult = @file_put_contents($this->logFile, $text, FILE_APPEND | LOCK_EX);
             if ($writeResult === false) {
                 $error = error_get_last();
@@ -163,7 +165,7 @@ class FileTarget extends Target
                     continue;
                 }
                 $newFile = $this->logFile . '.' . ($i + 1);
-                $this->rotateByCopy ? $this->rotateByCopy($rotateFile, $newFile) : $this->rotateByRename($rotateFile, $newFile);
+                $this->rotateByCopy($rotateFile, $newFile);
                 if ($i === 0) {
                     $this->clearLogFile($rotateFile);
                 }
@@ -194,15 +196,5 @@ class FileTarget extends Target
         if ($this->fileMode !== null) {
             @chmod($newFile, $this->fileMode);
         }
-    }
-
-    /**
-     * Renames rotated file into new file
-     * @param string $rotateFile
-     * @param string $newFile
-     */
-    private function rotateByRename($rotateFile, $newFile)
-    {
-        @rename($rotateFile, $newFile);
     }
 }